Implemented User Class, State handling using LoginManager, and other missing requirements.

+++ flaskr/backend.py
@@ -1,6 +1,5 @@
 from google.cloud import storage
 import json
-
 # TODO(Project 1): Implement Backend according to the requirements.
 class Backend:
 
@@ -42,6 +41,8 @@ class Backend:
         storage_client = storage.Client()
         bucket = storage_client.bucket("userpass")
         blob = bucket.blob(username)
+        if blob.exists():
+            return False
         with blob.open("w") as f:
             f.write(json.dumps(new_user))
         return storage.Blob(bucket=bucket, name=username).exists(storage_client)
